* check if deploy_to is a directory and raise exception if not  * silently skip if no composer.phar is found

+++ php/providers/composer.rb
@@ -4,7 +4,12 @@ end
 
 action :install do
   deploy_to = new_resource.name
-  if ::File.exists?("#{deploy_to}/composer.phar")
+  if !::File.directory?(deploy_to) 
+    raise "#{deploy_to} is not a directory"
+  end
+  if !::File.exists?("#{deploy_to}/composer.phar")
+    Chef::Log.info("Could not find 'composer.phar' in #{deploy_to}: silently skipping.")
+  else
     script "install dependencies with composer" do
       interpreter "bash"
       cwd         deploy_to
